backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2

# Celery & Redis
celery[redis]==5.3.4
redis==4.6.0

# AI/ML
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# AI API Clients
google-generativeai>=0.8.0

# Telegram Bot
python-telegram-bot>=20.7,<21.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# OCR
pytesseract==0.3.10

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
structlog==23.2.0
python-dateutil==2.8.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
